Add Property Label and Fix TTL
I add property label (directly on the file) and fix the mistype rdf:label that still persist in the data-008.ttl file. Saved it as data_009.ttl . The newest json for rdf transform is in openRefine-v3.json, although it only consist the fix rdf:label, not the addition of property label.